Add 'API keys' to preferences
This will create an application    for the    user and display the Client ID
and Access Token. There's a reset button to regenerate keys. This is a
bit of a hack since there's no real relation between apps and users
right now, so the app needs to be looked up by the account's username,
which is less than ideal.